requirements.txt
pandas
numpy
requests
hmmlearn
scikit-learn
tensorflow
keras
backtrader
matplotlib
seaborn
yfinance
tqdm
python-dotenv
beautifulsoup4
scrapy
tweepy
praw
cybotrade>=1.5.0
plotly.io
plotly.graph_objs